refactor!: The `protocolVersion` in the response is now represented as an `Integer`.

+++ src/main/java/org/stellar/sdk/responses/sorobanrpc/GetLatestLedgerResponse.java
@@ -12,7 +12,7 @@ import lombok.Value;
 public class GetLatestLedgerResponse {
   String id;
 
-  Long protocolVersion;
+  Integer protocolVersion;
 
   Integer sequence;
 }
